Enable stricter warnings and fix them
Enable stricter warnings that are used by client projects like VTK-m, so that
they are cought sooner.

This will reduce the number of backported changes from VTK-m.

Move thirdparty includes into a separate directory and include them as "system"

+++ examples/serialization/test-serialization.cpp
@@ -12,9 +12,9 @@ struct Point
 struct PointVec
 {
         PointVec()              {}          // need this for de-serialization
-        PointVec(int dim):
+        PointVec(size_t dim) :
           coords(dim)           {}
-  int&  operator[](int i)       { return coords[i]; }
+  int&  operator[](size_t i)       { return coords[i]; }
   std::vector<int>  coords;
 };
